Modif de l'id en int

+++ src/Utilisateur.java
@@ -7,7 +7,7 @@ public class Utilisateur {
     /**
      * Id de l'utilisateur
      */
-    private double id;
+    private int id;
 
     /**
      * Pseudo de l'utilisateur
@@ -43,7 +43,7 @@ public class Utilisateur {
     /**
      * Default constructor
      */
-    public Utilisateur(double id,String pseudo, String email, String mdp, Roles role) {
+    public Utilisateur(int id,String pseudo, String email, String mdp, Roles role) {
         this.id = id;
         this.pseudo = pseudo;
         this.email = email;
@@ -56,7 +56,7 @@ public class Utilisateur {
      * Getter id
      * @return (double) id
      */
-    public double getId() {
+    public int getId() {
         return id;
     }
 
@@ -64,7 +64,7 @@ public class Utilisateur {
      * Setter id
      * @param id
      */
-    public void setId(double id) {
+    public void setId(int id) {
         this.id = id;
     }
